requirements.txt
absl-py==1.0.0
addict==2.4.0
antlr4-python3-runtime==4.8
appdirs==1.4.4
audioread==3.0.1
black==21.4b2
blessed==1.19.1
blessings==1.7
Brotli==1.0.9
cachetools==4.2.4
certifi==2022.6.15
cffi==1.17.1
charset-normalizer==2.0.11
click==8.0.3
cloudpickle==2.0.0
coloredlogs==14.0
contourpy==1.1.1
cycler==0.12.1
dataclasses==0.6
decorator==5.1.1
decord==0.6.0
docker-pycreds==0.4.0
dynaconf==3.1.7
einops==0.8.0
enlighten==1.10.2
evdev==1.6.1
fairscale==0.4.8
filelock==3.13.1
fonttools==4.34.4
fsspec==2023.12.2
future==0.18.2
gitdb==4.0.11
GitPython==3.1.40
google-auth==1.35.0
google-auth-oauthlib==0.4.6
gpustat==0.6.0
grpcio==1.51.3
huggingface-hub==0.26.2
humanfriendly==10.0
hydra-core==1.1.1
idna==3.3
imageio==2.35.1
importlib-metadata==7.0.0
importlib-resources==6.1.1
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.41.1
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.4
mdurl==0.1.2
mmaction2==1.2.0
mmcv==1.3.4
mmengine==0.10.5
mpmath==1.3.0
msgpack==1.1.0
mypy-extensions==1.0.0
networkx==3.1
numba==0.58.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py3==7.352.0
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.1.2
opencv-contrib-python==4.5.5.62
opencv-python==4.7.0.72
packaging==23.2
pandas==2.0.3
pathspec==0.12.1
Pillow==10.1.0
piqa==1.3.2
platformdirs==4.3.6
pooch==1.8.2
prefixed==0.7.0
protobuf==3.20.1
psutil==5.9.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.22
Pygments==2.18.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
resampy==0.4.3
rich==13.9.4
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
sentry-sdk==2.18.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
soundfile==0.12.1
soxr==0.3.7
style==1.1.0
sympy==1.12
tabulate==0.8.9
tbb==2021.8.0
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
termcolor==2.4.0
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2023.7.10
timm==0.6.12
toml==0.10.2
tomli==2.0.2
torch==1.8.1+cu111
torchaudio==0.8.1
torchsummary==1.5.1
torchvision==0.9.1+cu111
tqdm==4.62.3
triton==2.1.0
typed-ast==1.5.2
typing_extensions==4.12.2
tzdata==2023.3
update==0.0.1
urllib3==1.26.18
vibromaf==0.0.6
wandb==0.18.7
wcwidth==0.2.13
Werkzeug==3.0.3
yapf==0.40.2
zipp==3.19.2
